Final analysis + small refactoring
