+++ path_finder/__init__.py
@@ -1,0 +1,1 @@
+from .algorithm import PathFinder, Path
